requirements.txt
altair==4.2.0
APScheduler==3.6.3
attrs==22.2.0
blinker==1.5
boto3==1.20.17
botocore==1.23.17
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
dash==2.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
distlib==0.3.6
dnspython==2.1.0
entrypoints==0.4
et-xmlfile==1.1.0
filelock==3.9.0
Flask==2.0.2
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.8.0
google-auth-oauthlib==0.5.2
gspread==5.4.0
idna==3.3
importlib-metadata==6.0.0
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.17.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.31
numpy==1.21.4
oauthlib==3.2.0
openpyxl==3.0.9
packaging==22.0
pandas==1.3.4
Pillow==9.4.0
pipenv==2022.12.19
platformdirs==2.6.2
plotly==5.11.0
protobuf==3.20.3
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
Pygments==2.14.0
pymongo==4.0.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.12
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.0
requests-oauthlib==1.3.1
rich==13.0.1
rsa==4.8
s3transfer==0.5.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.16.0
tenacity==8.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.1
typing_extensions==4.4.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.7
validators==0.20.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
watchdog==2.2.1
Werkzeug==2.0.2
XlsxWriter==3.0.2
zipp==3.11.0